pyproject.toml
[tool.poetry]
name = "swebench-qa"
version = "0.1.0"
description = ""
authors = ["Aaditya Bhatia", "Gustavo Oliva", "Gopi Rajbahadur"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
aider-chat = "0.82.2"
pygithub = "^2.6.1"
pandas = "^2.2.3"
pyarrow = "^19.0.1"
gitpython = "^3.1.44"
giturlparse = "^0.12.0"
unidiff = "^0.7.5"
pexpect = "^4.9.0"
openpyxl = "^3.1.5"
markdown2 = "^2.5.3"
markdownify = "^1.1.0"
anthropic = "^0.49.0"
scikit-learn = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
